# swagger/doc.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "My Portfolio API",
    "description": "Complete API reference for the Go/Fiber-powered portfolio backend.\n\n## Authentication\n- **Admin routes** (`/admin/*`) require a session cookie obtained via `POST /admin/login`.\n- **Comment POST** requires an OAuth session cookie obtained via GitHub or Google OAuth flow.\n- All other routes are public.\n\n## Rate Limits\n- Global: 200 req/min per IP (static assets and health probes exempt).\n- `POST /contact`: configurable, default 5 req/min per IP.\n- `POST /comments`: configurable, default 20 req/min per IP.\n- `POST /admin/login`: 10 req per 15 min per IP (brute-force guard).\n\n## HTMX Partials\nMany endpoints return HTML fragments for HTMX consumption rather than full pages or JSON. Response descriptions note when this applies.",
    "version": "0.0.1",
    "license": { "name": "MIT" }
  },
  "servers": [{ "url": "/", "description": "Local / production server" }],
  "components": {
    "securitySchemes": {
      "adminSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session",
        "description": "Session cookie set after successful POST /admin/login."
      },
      "oauthSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "oauth_session",
        "description": "OAuth session cookie set after completing Google or GitHub OAuth flow."
      }
    },
    "schemas": {
      "ErrorText": { "type": "string", "description": "Plain-text error message." },
      "ErrorJSON": {
        "type": "object",
        "properties": {
          "error":       { "type": "string" },
          "retry_after": { "type": "integer" }
        }
      },
      "OwnerForm": {
        "type": "object",
        "properties": {
          "full_name": { "type": "string", "maxLength": 100 },
          "title":     { "type": "string", "maxLength": 150 },
          "tagline":   { "type": "string", "maxLength": 300 },
          "bio":       { "type": "string" },
          "email":     { "type": "string", "maxLength": 100 },
          "phone":     { "type": "string", "maxLength": 30 },
          "location":  { "type": "string", "maxLength": 100 }
        }
      },
...
      "PostForm": {
        "type": "object",
        "required": ["title"],
        "properties": {
          "title":             { "type": "string", "maxLength": 300 },
          "excerpt":           { "type": "string" },
          "content":           { "type": "string", "description": "Markdown content with optional raw HTML blocks." },
          "tags":              { "type": "string", "maxLength": 500, "description": "Comma-separated tags." },
          "status":            { "type": "string", "enum": ["draft", "published"], "default": "draft" },
          "sort_order":        { "type": "integer", "default": 0 },
          "thumbnail_file_id": { "type": "integer", "nullable": true }
        }
      },
...
    },
    "responses": {
      "HTML200":        { "description": "HTML page.", "content": { "text/html": { "schema": { "type": "string" } } } },
      "HTMLPartial200": { "description": "HTMX HTML fragment (not a full page).", "content": { "text/html": { "schema": { "type": "string" } } } },
...
      "Unauthorized":   { "description": "Not authenticated.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorJSON" } } } },
      "BadRequest":     { "description": "Validation error.", "content": { "text/plain": { "schema": { "$ref": "#/components/schemas/ErrorText" } } } },
      "NotFound":       { "description": "Resource not found.", "content": { "text/plain": { "schema": { "$ref": "#/components/schemas/ErrorText" } } } },
...
      "ServiceUnavailable": { "description": "Server overloaded or DB not ready.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorJSON" } } } }
    }
  },
...
  "paths": {
...
    "/resume": {
      "get": { "summary": "View resume PDF inline", "operationId": "viewResume", "tags": ["Portfolio"],
        "responses": { "200": { "description": "PDF (Content-Disposition: inline).", "content": { "application/pdf": {} } }, "404": { "$ref": "#/components/responses/NotFound" } } }
    },
...
    "/blog/more": {
      "get": { "summary": "Blog posts HTMX partial (load more)", "operationId": "blogPostsPartial", "tags": ["Blog"],
        "description": "Returns an HTML partial with the next page of blog post cards for HTMX infinite scroll. Protected by circuit breaker.",
        "parameters": [{ "name": "page", "in": "query", "schema": { "type": "integer", "default": 2, "minimum": 2 } }],
        "responses": { "200": { "$ref": "#/components/responses/HTMLPartial200" }, "503": { "$ref": "#/components/responses/ServiceUnavailable" } } }
    },
...
    "/admin/owner": {
      "get": { "summary": "Owner edit page", "operationId": "adminOwnerEditPage", "tags": ["Admin · Owner"],
        "security": [{ "adminSession": [] }],
        "responses": { "200": { "$ref": "#/components/responses/HTML200" }, "401": { "$ref": "#/components/responses/Unauthorized" } } },
      "put": { "summary": "Update owner profile", "operationId": "adminOwnerUpdate", "tags": ["Admin · Owner"],
        "security": [{ "adminSession": [] }],
        "requestBody": { "required": true, "content": { "application/x-www-form-urlencoded": { "schema": { "$ref": "#/components/schemas/OwnerForm" } } } },
        "responses": { "200": { "$ref": "#/components/responses/HTMLPartial200" }, "400": { "$ref": "#/components/responses/BadRequest" }, "401": { "$ref": "#/components/responses/Unauthorized" } } }
    },
...
    "/admin/posts": {
      "get": { "summary": "Blog posts admin page", "operationId": "adminPostsPage", "tags": ["Admin · Blog Posts"],
        "security": [{ "adminSession": [] }],
        "responses": { "200": { "$ref": "#/components/responses/HTML200" }, "401": { "$ref": "#/components/responses/Unauthorized" } } },
      "post": { "summary": "Create blog post", "operationId": "adminPostCreate", "tags": ["Admin · Blog Posts"],
        "description": "Slug is auto-generated from title. published_at is set automatically on first publish.",
        "security": [{ "adminSession": [] }],
        "requestBody": { "required": true, "content": { "application/x-www-form-urlencoded": { "schema": { "$ref": "#/components/schemas/PostForm" } } } },
        "responses": { "200": { "$ref": "#/components/responses/HTMLPartial200" }, "400": { "$ref": "#/components/responses/BadRequest" }, "401": { "$ref": "#/components/responses/Unauthorized" } } }
    },
...
  }
}